refactor : 매개변수 줄이기 위해 컴포넌트별 기능 세분화
